YouTube support
Note: Userscript youtube player does not load in GreaseMonkey, must use ViolentMonkey to access unsafe windows. Not sure if I'm doing something wrong there, but while I was able to pass onyoutubeplayerapi by injecting js code to the page and using window.event, I was not able to pass over the YT.Player class from the imported js to the player with unsafeWindow, global variables, or even trying to push it through an event.